+++ postgis/Makefile.in
@@ -83,6 +83,7 @@ PG_OBJS= \
 	lwgeom_box.o \
 	lwgeom_box3d.o \
 	lwgeom_geos.o \
+	lwgeom_geos_predicates.o \
 	lwgeom_geos_prepared.o \
 	lwgeom_geos_clean.o \
 	lwgeom_geos_relatematch.o \
